Split BookForm into BookCreateForm and BookUpdateForm
