+++ src/features/books/BookFormFields.tsx
@@ -1,26 +1,20 @@
 import {
-  ActionIcon,
   Checkbox,
-  Group,
   Loader,
   MultiSelect,
   NumberInput,
   Select,
   Stack,
-  Text,
   TextInput,
 } from "@mantine/core";
-import { useForm } from "@mantine/form";
-import { zodResolver } from "mantine-form-zod-resolver";
-import React, { ReactElement } from "react";
+import { UseFormReturnType } from "@mantine/form";
+import React from "react";
 import { z } from "zod";
-import { IconSearch } from "@tabler/icons-react";
 import { BookFormat, BookStore } from "../../generated/graphql-request";
 import { useAuthors } from "../../compoments/hooks/useAuthors";
 import { Author } from "./entity/Author";
 import { BOOK_FORMAT_VALUE, displayBookFormat } from "./entity/BookFormat";
 import { BOOK_STORE_VALUE, displayBookStore } from "./entity/BookStore";
-import { useIsbnLookup } from "./useIsbnLookup";
 
 export type BookFormValues = {
   title: string;
@@ -33,7 +27,7 @@ export type BookFormValues = {
   store: BookStore;
 };
 
-const bookFormSchema = z.object({
+export const bookFormSchema = z.object({
   title: z.string().min(1),
   authors: z.array(z.object({ id: z.string(), name: z.string() })).nonempty(),
   isbn: z.string().regex(/(^$|^(\d-?){12}\d$)/),
@@ -44,92 +38,26 @@ const bookFormSchema = z.object({
   owned: z.boolean().default(false),
 });
 
-type BookFormProps = {
-  onSubmit: (
-    values: BookFormValues,
-    event: React.SyntheticEvent<HTMLFormElement> | undefined,
-  ) => void;
-  initialValues: BookFormValues;
-  enableIsbnLookup?: boolean;
+type BookFormFieldsProps = {
+  form: UseFormReturnType<BookFormValues>;
+  extraFields?: React.ReactNode;
 };
 
-type BookFormReturn = {
-  form: ReactElement;
-  submitForm: React.EventHandler<React.SyntheticEvent<HTMLFormElement>>;
-};
-
-export const useBookForm = (props: BookFormProps): BookFormReturn => {
-  const form = useForm({
-    initialValues: props.initialValues,
-    validate: zodResolver(bookFormSchema),
-    validateInputOnBlur: true,
-  });
-
+export const BookFormFields: React.FC<BookFormFieldsProps> = ({
+  form,
+  extraFields,
+}) => {
   const { data, isLoading, error } = useAuthors();
-  const { state: isbnLookupState, lookup: lookupIsbn } = useIsbnLookup();
-
-  const handleIsbnLookup = async () => {
-    const result = await lookupIsbn(form.values.isbn);
-    if (result == null || data == null) return;
-    form.setFieldValue("title", result.title);
-    const matched = Array.from(
-      new Map(
-        result.authorNames
-          .map((name) =>
-            data.authors.find(
-              (a) => a.name.trim().toLowerCase() === name.trim().toLowerCase(),
-            ),
-          )
-          .filter((a): a is Author => a !== undefined)
-          .map((a) => [a.id, a]),
-      ).values(),
-    );
-    if (matched.length > 0) {
-      form.setFieldValue("authors", matched);
-    }
-  };
-
-  const submitForm = form.onSubmit(props.onSubmit);
 
   if (error) {
-    return { form: <div>{JSON.stringify(error)}</div>, submitForm };
+    return <div>{JSON.stringify(error)}</div>;
   }
 
   if (isLoading || data == null) {
-    return { form: <Loader />, submitForm };
+    return <Loader />;
   }
 
-  const isbnField = props.enableIsbnLookup ? (
-    <Stack gap={0}>
-      <Group align="flex-end" gap="xs">
-        <TextInput
-          label="ISBN"
-          style={{ flex: 1 }}
-          {...form.getInputProps("isbn")}
-        />
-        <ActionIcon
-          onClick={() => {
-            void handleIsbnLookup();
-          }}
-          loading={isbnLookupState.status === "loading"}
-          size="lg"
-          variant="default"
-          aria-label="自動入力"
-        >
-          <IconSearch size={16} />
-        </ActionIcon>
-      </Group>
-      {isbnLookupState.status === "error" && (
-        <Text size="xs" c="red" mt={4} role="alert">
-          {isbnLookupState.message}
-        </Text>
-      )}
-    </Stack>
-  ) : (
-    <TextInput label="ISBN" {...form.getInputProps("isbn")} />
-  );
-
-  const formElement = (
+  return (
     <Stack>
       <TextInput label="書名" {...form.getInputProps("title")} />
       <MultiSelect
@@ -168,7 +96,9 @@ export const useBookForm = (props: BookFormProps): BookFormReturn => {
         }))}
       />
       <NumberInput label="優先度" {...form.getInputProps("priority")} />
-      {isbnField}
+      {extraFields ?? (
+        <TextInput label="ISBN" {...form.getInputProps("isbn")} />
+      )}
       <Checkbox
         label="既読"
         {...form.getInputProps("read", { type: "checkbox" })}
@@ -179,6 +109,4 @@ export const useBookForm = (props: BookFormProps): BookFormReturn => {
       />
     </Stack>
   );
-
-  return { form: formElement, submitForm };
 };
